// src/modules/core/@types/gUser.d.ts
export interface facebookUser {
    "userID": "-",
    "expiresIn": 0,
    "accessToken": "-",
    "signedRequest": "-",
    "graphDomain": "facebook",
    "grantedScopes": "-",
    "data_access_expiration_time": 0,
    "id": "-",
    "first_name": "-",
    "last_name": "-",
    "name": "- -",
    "name_format": "{first} {last}",
    "picture": {
      "data": {
        "height": 50,
        "is_silhouette": false,
        "url": "-",
        "width": 50
      }
    },
    "short_name": "-",
    "email": "-"
  }


  export interface googleUser {
    "id": "-",
    "email": "-",
    "verified_email": true,
    "name": "-",
    "given_name": "-",
    "family_name": "-",
    "picture": "-"
  }